backend/monthly_taxation/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0

# Authentication and security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and utilities
email-validator==2.1.0
python-dateutil==2.8.2

# Async and concurrency
asyncio==3.4.3

# Logging and monitoring
structlog==23.2.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1 